stream_proxy/celery_worker/requirements.txt
celery
redis
flower
qdrant-client
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-llms-ollama
rich
python-dotenv